more fixes from PR
