+++ dockstore-webservice/src/main/java/io/dockstore/webservice/helpers/GitHelper.java
@@ -1,13 +1,11 @@
 package io.dockstore.webservice.helpers;
 
+import java.util.Optional;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import io.dockstore.webservice.CustomWebApplicationException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import static io.dockstore.webservice.Constants.LAMBDA_FAILURE;
 
 /**
  * A helper class for handling git related problems
@@ -22,21 +20,19 @@ public final class GitHelper {
     }
 
     /**
-     * Parse git references (ex. refs/heads/feature/foobar) and returns a reference name
+     * Parse git references (ex. refs/heads/feature/foobar) and returns an optional reference name
      * @param gitReference Git Reference of the form refs/tags/name or refs/heads/name
-     * @return Git reference name
+     * @return Optional Git reference name
      */
-    public static String parseGitHubReference(String gitReference) {
+    public static Optional<String> parseGitHubReference(String gitReference) {
         // Match the github reference (ex. refs/heads/feature/foobar or refs/tags/1.0)
         Pattern pattern = Pattern.compile("^refs/(tags|heads)/([a-zA-Z0-9]+([./_-]?[a-zA-Z0-9]+)*)$");
         Matcher matcher = pattern.matcher(gitReference);
 
         if (!matcher.find()) {
-            String msg = "Reference " + gitReference + " is not of the valid form";
-            LOG.error(msg);
-            throw new CustomWebApplicationException(msg, LAMBDA_FAILURE);
+            return Optional.empty();
         }
         String gitBranchName = matcher.group(2);
-        return gitBranchName;
+        return Optional.of(gitBranchName);
     }
 }
