+++ dockstore-webservice/src/main/resources/swagger.yaml
@@ -4697,8 +4697,6 @@ paths:
         \ versions that match in all workflows that share the same repository."
       description: ""
       operationId: "handleGitHubBranchDeletion"
-      consumes:
-      - "application/x-www-form-urlencoded"
       produces:
       - "application/json"
       parameters:
